Allow multiple presence

+++ daemon/icd-dbus.h
@@ -24,7 +24,7 @@
 #define ICD_POINTER_TO_INT64(p) ((int64_t)(intptr_t)(p))
 
 icDbus* icd_dbus_get_object();
-int icd_dbus_client_list_get_info(void *handle, unsigned int *signal_number,
+int icd_dbus_client_list_get_info(void *handle, int64_t *signal_number,
 		gchar **bus_name);
 int icd_dbus_emit_signal(const char *dest, const char *signal_name,
 		GVariant *value);
